santexpo2024-dev/requirements.txt
langchain
faiss-cpu
openai
# gradio
python-dotenv
transformers
langchain-community
Flask
sentence-transformers
pypdf
torch
#pour pourvoir utiliser le GPU pour l'embedding, exécutez cette ligne de code :
#pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121